# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Vega API
  description: REST API for Vega, an AI agent orchestration framework.
  version: "1.0.0"
servers:
  - url: https://your-vega.example.com
    description: Production
paths:
...
  /api/v1/files/read:
    get:
      tags: [Files]
      summary: Read file content
      operationId: readFile
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Relative path under workspace
      responses:
        "200":
          description: File content (text as UTF-8, binary as base64)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
...
    FileContentResponse:
      type: object
      properties:
        path:
          type: string
        content_type:
          type: string
        content:
          type: string
          description: UTF-8 text or base64-encoded binary
        encoding:
          type: string
          enum: [utf-8, base64]
        size:
          type: integer
          format: int64
...